refactor: move randomColor to separate file
and add some tests.
