+++ src/utils/randomColor.ts
@@ -1,0 +1,4 @@
+export const randomColor = (): string =>
+	`#${Math.floor(Math.random() * 16777215)
+		.toString(16)
+		.padEnd(6, '0')}`
